Studied test driven development(TDD) and unittest method

+++ unit_tests/test_driven_development/test_driven_development.py
@@ -1,10 +1,14 @@
 # TDD - Test Driven Development
 # Разработка приложений, управляемая тестами
 # В начале пишем тесты
+from random import choice
 
 
 # Приветствовать
 def greet(name, is_enemy):
+    # Утверждаем, что если is_enemy не bool объект
+    if not isinstance(is_enemy, bool):
+        raise ValueError('is_enemy must be a boolean type')
     if is_enemy:
         return f'Hello {name}! I will kill you, bastard!'
     else:
@@ -17,3 +21,16 @@ def eat_burgers(number):
         return 'Oh! I overate!'
     else:
         return 'Mmm! That was excellent!'
+
+
+# Летать
+def can_fly(name):
+    if name == 'Batman':
+        return True
+    else:
+        return False
+
+
+# Получить оружие
+def get_arsenal():
+    return choice(('knife', 'handgun', 'machine gun'))
